backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
sqlalchemy==2.0.27
psycopg2-binary>=2.9.10
redis==5.0.1
numpy>=1.26.4
pandas>=2.2.0
scipy>=1.13.0
scikit-learn>=1.4.0
hmmlearn==0.3.3
statsmodels==0.14.6
python-dotenv==1.0.0
pydantic>=2.9.0
web3==6.15.0
httpx==0.26.0
aiohttp>=3.9.0
apify-client>=1.8.1
pytest==8.1.1
pytest-asyncio==0.23.5
hypothesis==6.100.1
torch>=2.2.0
kagglehub>=0.3.0
langchain>=0.3.27
langchain-groq>=0.2.4
